requirements.txt
streamlit>=1.32.0
pandas>=2.1.0
plotly>=5.18.0
gspread>=6.0.0
google-auth>=2.25.0
google-auth-oauthlib>=1.2.0
openpyxl>=3.1.2
